backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pillow==10.4.0
pytesseract==0.3.13
openai==1.40.3
supabase==2.8.0
psycopg[binary,pool]==3.2.3
pydantic==2.7.4
pydantic-settings==2.3.3
rich==13.8.0
typer==0.12.3
python-dotenv==1.0.1
pgvector==0.3.2
httpx==0.27.0